Standardize getattr->getitem with a mixin.
